# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Social Seller Toolkit API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Caption {
  id: number;
  text: string;
  niche: string;
  tone: string;
  platform: string;
  type: string;
  category: string;
  usageCount: number;
  isSaved: boolean;
  createdAt: string;
}

export interface CaptionInput {
  text: string;
  niche: string;
  tone: string;
  platform: string;
  type: string;
  category: string;
}

export interface CaptionListResponse {
  captions: Caption[];
  total: number;
  page: number;
  totalPages: number;
}

export interface NicheStat {
  niche: string;
  count: number;
}

export interface CaptionGenerateInput {
  productName: string;
  niche: string;
  tone: string;
  promotionGoal: string;
  platforms: string[];
}

export interface GeneratedCaption {
  platform: string;
  text: string;
}

export interface GeneratedCaptions {
  captions: GeneratedCaption[];
}

export interface SaveCaptionInput {
  save: boolean;
}

export interface SavedCaption {
  id: number;
  captionId: number;
  createdAt: string;
}

export interface BroadcastTemplate {
  id: number;
  title: string;
  message: string;
  category: string;
  niche: string;
  usageCount: number;
  createdAt: string;
}

export interface BroadcastInput {
  title: string;
  message: string;
  category: string;
  niche: string;
}

export interface BroadcastGenerateInput {
  niche: string;
  promotionType: string;
  audience: string;
  /** @nullable */
  discountPercentage?: number | null;
  productName: string;
  businessName: string;
}

export interface GeneratedBroadcast {
  message: string;
  subject: string;
}

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  niche: string;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  createdAt: string;
}

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  niche: string;
  imageUrl?: string;
  inStock?: boolean;
}

export interface ProductUpdate {
  title?: string;
  description?: string;
  price?: number;
  niche?: string;
  imageUrl?: string;
  inStock?: boolean;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  confirmed: 'confirmed',
  payment_received: 'payment_received',
  processing: 'processing',
  shipped: 'shipped',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  total: number;
  /** @nullable */
  trackingNumber?: string | null;
  /** @nullable */
  courier?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
}

export type OrderDetailStatus = typeof OrderDetailStatus[keyof typeof OrderDetailStatus];


export const OrderDetailStatus = {
  confirmed: 'confirmed',
  payment_received: 'payment_received',
  processing: 'processing',
  shipped: 'shipped',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderDetailPaymentStatus = typeof OrderDetailPaymentStatus[keyof typeof OrderDetailPaymentStatus];


export const OrderDetailPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  productTitle: string;
  quantity: number;
  price: number;
}

export interface OrderDetail {
  id: number;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  status: OrderDetailStatus;
  paymentStatus: OrderDetailPaymentStatus;
  total: number;
  /** @nullable */
  trackingNumber?: string | null;
  /** @nullable */
  courier?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  items: OrderItemInput[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  confirmed: 'confirmed',
  payment_received: 'payment_received',
  processing: 'processing',
  shipped: 'shipped',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  trackingNumber?: string;
  courier?: string;
  estimatedDelivery?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface OrderSummary {
  totalOrders: number;
  totalRevenue: number;
  activeOrders: number;
  deliveredOrders: number;
  cancelledOrders: number;
  pendingPayments: number;
  statusBreakdown: StatusCount[];
}

export interface WishlistItem {
  id: number;
  product: Product;
  createdAt: string;
}

export interface WishlistInput {
  productId: number;
}

export type LoyaltyTier = typeof LoyaltyTier[keyof typeof LoyaltyTier];


export const LoyaltyTier = {
  bronze: 'bronze',
  silver: 'silver',
  gold: 'gold',
  platinum: 'platinum',
} as const;

export interface Loyalty {
  points: number;
  tier: LoyaltyTier;
  nextTierPoints: number;
  totalSpent: number;
  discountEarned: number;
  referralCode: string;
}

export interface DayStat {
  date: string;
  revenue: number;
  orders: number;
}

export interface AnalyticsOverview {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  captionsUsed: number;
  broadcastsSent: number;
  topNiches: NicheStat[];
  revenueByDay: DayStat[];
  repeatPurchaseRate: number;
}

export type ListCaptionsParams = {
niche?: string;
tone?: string;
platform?: string;
type?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListBroadcastsParams = {
category?: string;
niche?: string;
search?: string;
};

export type ListProductsParams = {
search?: string;
niche?: string;
};

export type ListOrdersParams = {
status?: string;
/**
 * @nullable
 */
customerId?: number | null;
};