# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Social Seller Toolkit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /captions:
    get:
      operationId: listCaptions
      tags: [captions]
      summary: List captions with filters
      parameters:
        - name: niche
          in: query
          schema:
            type: string
        - name: tone
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of captions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptionListResponse"
    post:
      operationId: createCaption
      tags: [captions]
      summary: Create a new caption template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caption"
...
  /captions/generate:
    post:
      operationId: generateCaption
      tags: [captions]
      summary: AI-generate captions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptionGenerateInput"
      responses:
        "200":
          description: Generated captions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedCaptions"
  /captions/{id}:
    get:
      operationId: getCaption
      tags: [captions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Caption
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caption"
        "404":
          description: Not found
  /captions/{id}/save:
    post:
      operationId: saveCaption
      tags: [captions]
      summary: Save or unsave a caption
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCaptionInput"
      responses:
        "200":
          description: Save status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedCaption"
...
  /broadcasts:
    get:
      operationId: listBroadcasts
      tags: [broadcasts]
      summary: List broadcast templates
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: niche
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of broadcast templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BroadcastTemplate"
    post:
      operationId: createBroadcast
      tags: [broadcasts]
      summary: Create broadcast template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastTemplate"
  /broadcasts/generate:
    post:
      operationId: generateBroadcast
      tags: [broadcasts]
      summary: AI-generate broadcast message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastGenerateInput"
      responses:
        "200":
          description: Generated broadcast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedBroadcast"
  /broadcasts/{id}:
    get:
      operationId: getBroadcast
      tags: [broadcasts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Broadcast template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastTemplate"
        "404":
          description: Not found
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Caption:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        niche:
          type: string
        tone:
          type: string
        platform:
          type: string
        type:
          type: string
        category:
          type: string
        usageCount:
          type: integer
        isSaved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - text
        - niche
        - tone
        - platform
        - type
        - category
        - usageCount
        - isSaved
        - createdAt
    CaptionInput:
      type: object
      properties:
        text:
          type: string
        niche:
          type: string
        tone:
          type: string
        platform:
          type: string
        type:
          type: string
        category:
          type: string
      required:
        - text
        - niche
        - tone
        - platform
        - type
        - category
    CaptionListResponse:
      type: object
      properties:
        captions:
          type: array
          items:
            $ref: "#/components/schemas/Caption"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - captions
        - total
        - page
        - totalPages
...
    CaptionGenerateInput:
      type: object
      properties:
        productName:
          type: string
        niche:
          type: string
        tone:
          type: string
        promotionGoal:
          type: string
        platforms:
          type: array
          items:
            type: string
      required:
        - productName
        - niche
        - tone
        - promotionGoal
        - platforms
    GeneratedCaptions:
      type: object
      properties:
        captions:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedCaption"
      required:
        - captions
    GeneratedCaption:
      type: object
      properties:
        platform:
          type: string
        text:
          type: string
      required:
        - platform
        - text
    SaveCaptionInput:
      type: object
      properties:
        save:
          type: boolean
      required:
        - save
    SavedCaption:
      type: object
      properties:
        id:
          type: integer
        captionId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - captionId
        - createdAt
    BroadcastTemplate:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        category:
          type: string
        niche:
          type: string
        usageCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - category
        - niche
        - usageCount
        - createdAt
    BroadcastInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        category:
          type: string
        niche:
          type: string
      required:
        - title
        - message
        - category
        - niche
    BroadcastGenerateInput:
      type: object
      properties:
        niche:
          type: string
        promotionType:
          type: string
        audience:
          type: string
        discountPercentage:
          type: ["integer", "null"]
        productName:
          type: string
        businessName:
          type: string
      required:
        - niche
        - promotionType
        - audience
        - productName
        - businessName
    GeneratedBroadcast:
      type: object
      properties:
        message:
          type: string
        subject:
          type: string
      required:
        - message
        - subject
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [confirmed, payment_received, processing, shipped, out_for_delivery, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        total:
          type: number
        trackingNumber:
          type: ["string", "null"]
        courier:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - customerName
        - customerEmail
        - status
        - paymentStatus
        - total
        - createdAt
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [confirmed, payment_received, processing, shipped, out_for_delivery, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        total:
          type: number
        trackingNumber:
          type: ["string", "null"]
        courier:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - customerName
        - customerEmail
        - status
        - paymentStatus
        - total
        - items
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productTitle
        - quantity
        - price
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - customerEmail
        - items
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, payment_received, processing, shipped, out_for_delivery, delivered, cancelled]
        trackingNumber:
          type: string
        courier:
          type: string
        estimatedDelivery:
          type: string
      required:
        - status
...