# lib/api-zod/src/generated/types/generatedCaptions.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Social Seller Toolkit API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedCaption } from './generatedCaption';

export interface GeneratedCaptions {
  captions: GeneratedCaption[];
}